Refactor: todolist 작성, 수정 상태에서 다른 행동 불가

+++ src/components/Practice/lifecoding/advanced/Control.jsx
@@ -1,25 +1,45 @@
 import { useStore } from "./storeContext";
+import { useState } from "react";
 
 function Control() {
   const store = useStore();
+  const state = store.getState();
+  const [isDisabled, setIsDisabled] = useState(
+    state.mode === "read" ? false : true
+  );
   const handleCreateClick = () => {
-    store.dispatch({ type: "CREATE" });
+    if (!isDisabled) {
+      store.dispatch({ type: "CREATE" });
+    }
   };
   const handleModifyClick = () => {
     const state = store.getState();
-    if (state.contents.length > 0) {
+
+    if (!isDisabled && state.contents.length > 0) {
       store.dispatch({ type: "MODIFY" });
     }
   };
   const handleDeleteClick = () => {
-    store.dispatch({ type: "DELETE" });
+    if (!isDisabled) {
+      store.dispatch({ type: "DELETE" });
+    }
   };
+
+  // change mode
+  const handleModeChange = () => {
+    const newState = store.getState();
+    setIsDisabled(newState.mode === "read" ? false : true);
+  };
+
+  store.subscribe(handleModeChange);
 
   return (
     <ul className="control-container">
       <li>
         <input
           type={"button"}
+          disabled={isDisabled}
+          className={isDisabled ? "disabled" : ""}
           name={"create"}
           value={"작성"}
           onClick={handleCreateClick}
@@ -28,6 +48,8 @@ function Control() {
       <li>
         <input
           type={"button"}
+          disabled={isDisabled}
+          className={isDisabled ? "disabled" : ""}
           name={"modify"}
           value={"수정"}
           onClick={handleModifyClick}
@@ -36,6 +58,8 @@ function Control() {
       <li>
         <input
           type={"button"}
+          disabled={isDisabled}
+          className={isDisabled ? "disabled" : ""}
           name={"delete"}
           value={"삭제"}
           onClick={handleDeleteClick}
